api/requirements.txt
fastapi
uvicorn
torch
torchvision
boto3
pillow
python-multipart
